% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sir_timevary_generator.R
\name{make_sir_timevary_generator}
\alias{make_sir_timevary_generator}
\title{Discrete-time deterministic SIR with stepwise beta(t) and Poisson observation}
\usage{
make_sir_timevary_generator()
}
\description{
Discrete-time deterministic SIR with stepwise beta(t) and Poisson observation
}
