% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt_constant_wrapper.R
\name{rt_estimate_constant}
\alias{rt_estimate_constant}
\title{Simple V1 wrapper: deterministic SIR Rt(t), estimating beta & gamma (no priors)}
\usage{
rt_estimate_constant(incidence, N, I0)
}
\description{
Simple V1 wrapper: deterministic SIR Rt(t), estimating beta & gamma (no priors)
}
