% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_manysteps.R
\name{sir_manysteps}
\alias{sir_manysteps}
\title{Simulate SIR Model with Multiple Time-Varying Transmission Rates}
\usage{
sir_manysteps(
  N,
  I0,
  gamma,
  time_points,
  beta_values,
  end_time,
  time_step = 0.25
)
}
\arguments{
\item{N}{Total population size}

\item{I0}{Initial number of infected individuals}

\item{gamma}{Recovery rate (1/infectious period)}

\item{time_points}{Vector of time points where beta changes}

\item{beta_values}{Vector of beta values corresponding to time_points}

\item{end_time}{End time for simulation}

\item{time_step}{Time step for output (default: 0.25)}
}
\value{
A list containing:
\item{times}{Time points}
\item{S}{Susceptible counts over time}
\item{I}{Infected counts over time}
\item{R}{Recovered counts over time}
\item{beta}{Transmission rate over time}
\item{parameters}{List of parameters used}
}
\description{
Simulate SIR Model with Multiple Time-Varying Transmission Rates
}
