% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_onestep.R
\name{sir_onestep}
\alias{sir_onestep}
\title{Add together two numbers}
\usage{
sir_onestep(
  N,
  I0,
  gamma,
  beta1,
  beta2,
  change_time,
  times = seq(0, 100, by = 1)
)
}
\arguments{
\item{N}{Total population}
}
\description{
Add together two numbers
}
