% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_twostep.R
\name{sir_twostep}
\alias{sir_twostep}
\title{Simulate SIR Model with Two Time-Varying Transmission Rates}
\usage{
sir_twostep(
  N,
  I0,
  gamma,
  beta1,
  beta2,
  beta3,
  change_time1,
  change_time2,
  end_time
)
}
\arguments{
\item{N}{Total population size}

\item{I0}{Initial number of infected individuals}

\item{gamma}{Recovery rate (1/infectious period)}

\item{beta1}{Transmission rate in first period}

\item{beta2}{Transmission rate in second period}

\item{beta3}{Transmission rate in third period}

\item{change_time1}{First change point}

\item{change_time2}{Second change point}

\item{end_time}{End time for simulation}
}
\value{
A list containing simulation results
}
\description{
Simulate SIR Model with Two Time-Varying Transmission Rates
}
